fix: move email service and all related classes into the application layer.
