--- src/app/Application/Mail/MailerInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Application\Mail;
-
-use Illuminate\Mail\Mailable;
-
-interface MailerInterface
-{
-    public function send(string $to, Mailable $mailable): bool;
-}
