+++ src/app/Application/Subscription/Emails/Mailers/MailerInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Application\Subscription\Emails\Mailers;
+
+use Illuminate\Mail\Mailable;
+
+interface MailerInterface
+{
+    public function send(string $email, Mailable $mailable): bool;
+}
