+++ src/app/Application/Subscription/Emails/Mails/Update/WeatherUpdateMailData.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Infrastructure\Subscription\Mails\Update;
+namespace App\Application\Subscription\Emails\Mails\Update;
 
 use App\Domain\Weather\ValueObjects\WeatherData;
