+++ src/app/Application/Subscription/Emails/Services/EmailService.php
@@ -1,24 +1,24 @@
 <?php
 
-namespace App\Infrastructure\Subscription\Services;
+namespace App\Application\Subscription\Emails\Services;
 
-use App\Application\Mail\MailerInterface;
-use App\Application\Subscription\Services\EmailServiceInterface;
+use App\Application\Subscription\Emails\Mails\Confirmation\ConfirmationMail;
+use App\Application\Subscription\Emails\Mails\Confirmation\ConfirmationMailData;
+use App\Application\Subscription\Emails\Mails\Update\WeatherUpdateMail;
+use App\Application\Subscription\Emails\Mails\Update\WeatherUpdateMailData;
+use App\Application\Subscription\Emails\EmailServiceInterface;
+use App\Application\Subscription\Emails\Mailers\MailerInterface;
 use App\Application\Subscription\Utils\Builders\SubscriptionLinkBuilderInterface;
 use App\Domain\Subscription\Entities\Subscription;
 use App\Domain\Weather\ValueObjects\WeatherData;
-use App\Infrastructure\Subscription\Mails\Confirmation\ConfirmationMail;
-use App\Infrastructure\Subscription\Mails\Confirmation\ConfirmationMailData;
-use App\Infrastructure\Subscription\Mails\Update\WeatherUpdateMail;
-use App\Infrastructure\Subscription\Mails\Update\WeatherUpdateMailData;
 use App\Infrastructure\Subscription\Utils\Links\extend\ConfirmationLink;
 use App\Infrastructure\Subscription\Utils\Links\extend\UnsubscribeLink;
 
 class EmailService implements EmailServiceInterface
 {
     public function __construct(
-        private MailerInterface $mailer,
-        private SubscriptionLinkBuilderInterface $urlBuilder
+        private readonly MailerInterface $mailer,
+        private readonly SubscriptionLinkBuilderInterface $urlBuilder
     ) {
     }
