--- src/app/Application/Subscription/Utils/Links/Base/LinkWithToken.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Application\Subscription\Utils\Links\Base;
-
-interface LinkWithToken
-{
-    public function getToken(): ?string;
-}
