+++ src/app/Application/Subscription/Utils/Links/SubscriptionLinkInterface.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace App\Application\Subscription\Utils\Links;
+namespace App\Infrastructure\Subscription\Utils\Links\Inter;
 
-use App\Application\Subscription\Utils\Links\Base\LinkWithEndpoint;
-use App\Application\Subscription\Utils\Links\Base\LinkWithToken;
+use App\Infrastructure\Subscription\Utils\Links\Inter\Base\LinkWithEndpoint;
+use App\Infrastructure\Subscription\Utils\Links\Inter\Base\LinkWithToken;
 
 interface SubscriptionLinkInterface extends LinkWithToken, LinkWithEndpoint
 {
