+++ src/app/Infrastructure/Subscription/Emails/Mailers/LaravelMailer.php
@@ -2,7 +2,7 @@
 
 namespace App\Infrastructure\Mail;
 
-use App\Application\Mail\MailerInterface;
+use App\Application\Email\Mailers\MailerInterface;
 use Illuminate\Mail\Mailable;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Mail;
@@ -10,17 +10,17 @@ use Illuminate\Support\Facades\Mail;
 class LaravelMailer implements MailerInterface
 {
     /**
-     * @param string $to
+     * @param string $email
      * @param Mailable $mailable
      * @return bool
      */
-    public function send(string $to, Mailable $mailable): bool
+    public function send(string $email, Mailable $mailable): bool
     {
         try {
-            Mail::to($to)->send($mailable);
+            Mail::to($email)->send($mailable);
             return true;
         } catch (\Throwable $e) {
-            Log::error("Failed to send email to {$to}: " . $e->getMessage());
+            Log::error("Failed to send email to {$email}: " . $e->getMessage());
             return false;
         }
     }
