+++ src/app/Infrastructure/Subscription/Token/Generator/TokenGenerator.php
